add readme to thumbdrive improve error handling for worker

+++ desktop/src/main/java/de/felixnuesse/desktop/Main.kt
@@ -48,6 +48,11 @@ class AesTool : CliktCommand() {
             target
         }
 
+        if(targetFile.exists()) {
+            System.err.println("Target file exists, aborting!")
+            return
+        }
+
         if(encrypt) {
             Crypto().aesEncrypt(FileInputStream(source), FileOutputStream(targetFile), password.toCharArray())
         } else {
